Handle architecture mismatch to avoid crashing host
Fixes: https://github.com/wixtoolset/issues/issues/6636

MSBuild typically handles exceptions thrown directly by tasks and reports them as errors. WixToolTask.ExecuteTool, however, is forking its execution onto another thread and this thread is throwing when invoking Heat directly in-proc. This is crashing its host. This is #15 of all crashes in Visual Studio 17.3.

Handle the error and give a helpful message instead of crashing.