Revert the mergemod.cub to pre-ARM64 builds to fix MSM validation
When introducing ARM64 support into the Windows Installer, Microsoft broke
the ICE CUBe files in various ways. To minimize the impact of the breakage
move the mergemod.cub file back to pre-ARM64 support. Validating ARM64 Merge
Modules is not likely to work but that option is better than the regression.

Fixes wixtoolset/issues#8065